[PAY-572] Add animation to Show Advanced toggle
