+++ packages/web/src/pages/audio-rewards-page/components/WalletManagementTile.module.css
@@ -99,28 +99,6 @@
   align-self: flex-start;
 }
 
-.toggleCollapsedButton {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: var(--font-bold);
-  font-size: var(--font-s);
-  color: var(--neutral-light-4);
-  padding: 8px 24px;
-  gap: 5px;
-}
-.toggleCollapsedButton:hover {
-  cursor: pointer;
-}
-.toggleCollapsedButton:hover span {
-  color: var(--secondary);
-  transition: color 0.07s ease-out;
-}
-.toggleCollapsedButton:hover svg path {
-  fill: var(--secondary);
-  transition: fill 0.07s ease-out;
-}
-
 .advancedToggle {
   width: 100%;
   border-top: 1px solid var(--neutral-light-8);
@@ -133,6 +111,7 @@
   align-items: center;
   width: 100%;
   padding-top: 16px;
+  padding-bottom: 24px;
   gap: 16px;
   border-top: 1px solid var(--neutral-light-8);
 }
