+++ packages/web/src/pages/audio-rewards-page/components/WalletManagementTile.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react'
+import { useCallback } from 'react'
 
 import {
   Nullable,
@@ -13,8 +13,6 @@ import BN from 'bn.js'
 import cn from 'classnames'
 import { useDispatch, useSelector } from 'react-redux'
 
-import { ReactComponent as IconCaretDownLine } from 'assets/img/iconCaretDownLine.svg'
-import { ReactComponent as IconCaretUpLine } from 'assets/img/iconCaretUpLine.svg'
 import { ReactComponent as IconReceive } from 'assets/img/iconReceive.svg'
 import { ReactComponent as IconSend } from 'assets/img/iconSend.svg'
 import { ReactComponent as IconSettings } from 'assets/img/iconSettings.svg'
@@ -22,6 +20,7 @@ import IconGoldBadge from 'assets/img/tokenBadgeGold40@2x.png'
 import { useModalState } from 'common/hooks/useModalState'
 import { CoinbasePayButtonCustom } from 'components/coinbase-pay-button'
 import MobileConnectWalletsDrawer from 'components/mobile-connect-wallets-drawer/MobileConnectWalletsDrawer'
+import { ToggleCollapseButton } from 'components/toggle-collapse-button/ToggleCollapseButton'
 import { isMobile } from 'utils/clientUtil'
 
 import TokenHoverTooltip from './TokenHoverTooltip'
@@ -131,35 +130,6 @@ const AdvancedWalletActions = () => {
   )
 }
 
-const ToggleCollapseButton = ({
-  className,
-  showText,
-  hideText,
-  children
-}: {
-  className?: string
-  showText: string
-  hideText: string
-  children: React.ReactNode
-}) => {
-  const [isCollapsed, setIsCollapsed] = useState(true)
-  const handleToggle = useCallback(() => {
-    setIsCollapsed((isCollapsed) => !isCollapsed)
-  }, [setIsCollapsed])
-  return (
-    <>
-      {isCollapsed ? null : children}
-      <div
-        className={cn(styles.toggleCollapsedButton, className)}
-        onClick={handleToggle}
-      >
-        <span>{isCollapsed ? showText : hideText}</span>
-        {isCollapsed ? <IconCaretDownLine /> : <IconCaretUpLine />}
-      </div>
-    </>
-  )
-}
-
 export const WalletManagementTile = () => {
   const totalBalance: Nullable<BNWei> =
     useSelector(getAccountTotalBalance) ?? null
@@ -216,7 +186,7 @@ export const WalletManagementTile = () => {
         </div>
         <CoinbasePayButtonCustom onClick={onBuyAudioClicked} />
         <ToggleCollapseButton
-          className={styles.advancedToggle}
+          toggleButtonClassName={styles.advancedToggle}
           showText={messages.showAdvanced}
           hideText={messages.hideAdvanced}
         >
